+++ CHANGELOG.md
@@ -7,8 +7,8 @@ previous user inputs.
 
 ## Version 2.0.0 - *Upcoming*
 
-This version includes a total rewrite of the codebase and improvements to the
-`scan` command.
+This version includes a total rewrite of the codebase and improvements to
+all commands.
 
 ### Breaking Changes
 
@@ -16,6 +16,8 @@ This version includes a total rewrite of the codebase and improvements to the
   - The program will now **panic** if this keyvalue is not present or holds the
     wrong number.
   - The current `images.slop` version is now `1`.
+- `build` now needs the `sounds.slop` reference file to compare sound files and
+  will **panic** otherwise.
 
 #### Migrating From `1.0.0` to `2.0.0`
 
@@ -23,7 +25,9 @@ Simply run the `gen` command again in version [2.0.0].
 
 ### Added
 
+- `gen` now generates sound reference files.
 - `scan` now compares image sizes in addition to file names.
+- `scan` now compares sound files.
 
 ## Version 1.0.0 - 2023-08-08